[FEATURE] Flexible content dimensions (and localization)
This change implements content dimensions for nodes with fallback to
most specific matchers. This is a generic solution for the localization
problem that allows further dimensions in the future.

A new NodeDimension entity stores dimension values for each NodeData, so
a node variant can have multiple dimension values assigned (e.g. a set
of locales).

DimensionValues are also cached inside NodeData for optimized
reading speed.

Behat features provide scenarios to specify and test the localization
implementation against various use cases.

Needs Iadbe03a194141a49c8fea0e6715b07750f11c726 in TYPO3.Neos for the
functional tests and Behat features to pass because of the changes in
Context and ContextFactory.

Releases: master
Reviewed-on: https://review.typo3.org/27483
Tested-by: Karsten Dambekalns
Tested-by: Christian Mueller
Tested-by: Robert Lemke
Original-Commit-Hash: febd6bcda1f8ea24361302572fb01aefb04c0be7

+++ TYPO3.TYPO3CR/Classes/TYPO3/TYPO3CR/Domain/Model/Workspace.php
@@ -61,6 +61,12 @@ class Workspace {
 
 	/**
 	 * @Flow\Inject
+	 * @var \TYPO3\TYPO3CR\Domain\Service\ContextFactoryInterface
+	 */
+	protected $contextFactory;
+
+	/**
+	 * @Flow\Inject
 	 * @var ObjectManagerInterface
 	 */
 	protected $objectManager;
@@ -179,8 +185,11 @@ class Workspace {
 		if ($node->getPath() === '/') {
 			return;
 		}
-		$targetNode = $this->nodeDataRepository->findOneByIdentifier($node->getIdentifier(), $targetWorkspace);
-		if ($targetNode !== NULL) {
+
+		$targetNode = $this->findNodeDataInTargetWorkspace($node, $targetWorkspace);
+
+			// Only remove the node in the target workspace if it has the same dimensions, otherwise we want to keep it, since it's another node variant
+		if ($targetNode !== NULL && $targetNode->getDimensionValues() === $node->getDimensions()) {
 			if ($node->isRemoved() === TRUE) {
 				$this->nodeDataRepository->remove($targetNode);
 			} else {
@@ -228,6 +237,24 @@ class Workspace {
 			}
 			$baseWorkspace = $baseWorkspace->getBaseWorkspace();
 		}
+	}
+
+	/**
+	 * Returns the NodeData instance with the given identifier from the target workspace.
+	 * If no NodeData instance is found, NULL is returned.
+	 *
+	 * @param NodeInterface $node
+	 * @param Workspace $targetWorkspace
+	 * @return NodeData
+	 */
+	protected function findNodeDataInTargetWorkspace(NodeInterface $node, Workspace $targetWorkspace) {
+		$properties = $node->getContext()->getProperties();
+		$properties['workspaceName'] = $targetWorkspace->getName();
+		$targetWorkspaceContext = $this->contextFactory->create($properties);
+
+		$targetNodeInstance = $targetWorkspaceContext->getNodeByIdentifier($node->getIdentifier());
+		$targetNode = $targetNodeInstance !== NULL ? $targetNodeInstance->getNodeData() : NULL;
+		return $targetNode;
 	}
 
 	/**
